# perf(python): Pre-allocate size for the dictionary